SplashScreens for Applets have been included.
In addition the loading message that is embedded in the HTML file that triggers the Applet disappears when the game starts.

Now the text of the loading message has been translated to Spanish as well. It has been added to the language files.

Loading Image for GameStateLoading has been fixed: It was loaded using "new ImageIcon(path).getImage();". Now it uses the MultimediaManager.